requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
websockets==12.0
sounddevice==0.4.7
numpy==1.26.4
scipy==1.13.1
pydub==0.25.1
transformers==4.43.3
datasets==2.20.0
accelerate==0.33.0
torch
python-dotenv==1.0.1
torchaudio==2.8.0
openai>=1.0.0
